+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/RecallSpell.java
@@ -1,14 +1,21 @@
 package com.elmakers.mine.bukkit.plugins.magic.spells;
 
+import java.util.List;
+
+import org.bukkit.Bukkit;
+import org.bukkit.Chunk;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
+import org.bukkit.entity.Player;
+import org.bukkit.plugin.Plugin;
 
 import com.elmakers.mine.bukkit.effects.EffectUtils;
 import com.elmakers.mine.bukkit.effects.ParticleType;
 import com.elmakers.mine.bukkit.plugins.magic.Spell;
 import com.elmakers.mine.bukkit.plugins.magic.SpellResult;
+import com.elmakers.mine.bukkit.plugins.magic.wand.LostWand;
 import com.elmakers.mine.bukkit.utilities.borrowed.ConfigurationNode;
 
 public class RecallSpell extends Spell
@@ -16,6 +23,11 @@ public class RecallSpell extends Spell
 	public Location location;
 	public boolean isActive;
 	private int disableDistance = 5;
+
+	private static int MAX_RETRY_COUNT = 8;
+	private static int RETRY_INTERVAL = 10;
+	
+	private int retryCount = 0;
 
 	@Override
 	public SpellResult onCast(ConfigurationNode parameters) 
@@ -27,7 +39,20 @@ public class RecallSpell extends Spell
 		if (typeString.equals("spawn"))
 		{
 			castMessage("Returning you home");
-			getPlayer().teleport(tryFindPlaceToStand(getPlayer().getWorld().getSpawnLocation()));
+			tryTeleport(getPlayer().getWorld().getSpawnLocation());
+			return SpellResult.SUCCESS; 
+		}
+		if (typeString.equals("wand"))
+		{
+			List<LostWand> lostWands = mage.getLostWands();
+			Location wandLocation = lostWands.size() > 0 ? lostWands.get(0).getLocation() : null;
+			if (wandLocation == null)
+			{
+				sendMessage("No recorded lost wands for you. Sorry!");
+				return SpellResult.NO_TARGET;
+			}
+			
+			tryTeleport(wandLocation);
 			return SpellResult.SUCCESS; 
 		}
 		if (typeString.equals("death") || getYRotation() < -70  && autoResurrect)
@@ -35,10 +60,11 @@ public class RecallSpell extends Spell
 			Location deathLocation = mage.getLastDeathLocation();
 			if (deathLocation == null)
 			{
+				sendMessage("No recorded death location. Sorry!");
 				return SpellResult.NO_TARGET;
 			}
 			
-			getPlayer().teleport(tryFindPlaceToStand(deathLocation));
+			tryTeleport(deathLocation);
 			return SpellResult.SUCCESS; 
 		}
 		
@@ -47,7 +73,7 @@ public class RecallSpell extends Spell
 			if (!isActive && autoSpawn)
 			{
 				castMessage("Returning you home");
-				getPlayer().teleport(tryFindPlaceToStand(getPlayer().getWorld().getSpawnLocation()));
+				tryTeleport(getPlayer().getWorld().getSpawnLocation());
 			}
 			else
 			{
@@ -58,12 +84,12 @@ public class RecallSpell extends Spell
 				if (distanceSquared < disableDistance * disableDistance && autoSpawn)
 				{
 					castMessage("Returning you home");
-					getPlayer().teleport(tryFindPlaceToStand(getPlayer().getWorld().getSpawnLocation()));
+					tryTeleport(getPlayer().getWorld().getSpawnLocation());
 				}
 				else
 				{
 					castMessage("Returning you to your marker");
-					getPlayer().teleport(location);
+					tryTeleport(location);
 				}
 			}
 			return SpellResult.SUCCESS;
@@ -82,6 +108,32 @@ public class RecallSpell extends Spell
 		if (!isActive || location == null) return false;
 		isActive = false;
 		return true;
+	}
+	
+	protected void tryTeleport(final Location targetLocation) {
+		Chunk chunk = targetLocation.getBlock().getChunk();
+		if (!chunk.isLoaded()) {
+			chunk.load(true);
+			if (retryCount < MAX_RETRY_COUNT) {
+				Plugin plugin = controller.getPlugin();
+				final RecallSpell me = this;
+				Bukkit.getScheduler().scheduleSyncDelayedTask(plugin, new Runnable() {
+					public void run() {
+						me.tryTeleport(targetLocation);
+					}
+				}, RETRY_INTERVAL);
+				
+				return;
+			}
+		}
+		
+		Player player = getPlayer();
+		if (player != null) {
+			Location playerLocation = player.getLocation();
+			targetLocation.setYaw(playerLocation.getYaw());
+			targetLocation.setPitch(playerLocation.getPitch());
+			player.teleport(tryFindPlaceToStand(targetLocation));
+		}
 	}
 
 	protected SpellResult placeMarker(Block target)