+++ CHANGELOG.md
@@ -11,6 +11,7 @@
  - Add phase spell
  - Superblob will now replace more block types. Blob remains the same.
  - Some general fixes to the "find place to stand" code (ascend, descend, phase)
+ - Add "wand" recall type, some general recall improvements.
 
 ## 2.8.8